+++ http.go
@@ -1,9 +1,8 @@
 package sola
 
 import (
+	"net/http"
 	"sync"
-
-	"github.com/fvbock/endless"
 )
 
 var wg sync.WaitGroup
@@ -13,7 +12,7 @@ func Listen(addr string, g *Sola) {
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		endless.ListenAndServe(addr, g)
+		http.ListenAndServe(addr, g)
 	}()
 }
 
@@ -22,7 +21,7 @@ func ListenTLS(addr string, certFile string, keyFile string, g *Sola) {
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
-		endless.ListenAndServeTLS(addr, certFile, keyFile, g)
+		http.ListenAndServeTLS(addr, certFile, keyFile, g)
 	}()
 }