+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 + 新路由中间件省缺路径，仅指定方法
++ 由于 endless 无法在 Windows 下运行，单独给 linux 写平滑切换
 
 ### Features